pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "media-toolkit"
version = "0.2.0"
description = "Social media collection + general media download tool"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your@email.com"}
]
keywords = ["social-media", "scraper", "data-collection", "media-downloader", "yt-dlp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "pydantic>=2.0",
    "aiohttp>=3.9",
    "aiofiles>=23.0",
    "yt-dlp>=2024.1.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "typer>=0.9",
    "Pillow>=10.0",
    "python-frontmatter>=1.0",
    "streamlit",
    "cloudscraper",
    "fake-useragent",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
]

[project.scripts]
media-toolkit = "media_toolkit.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
